+++ projects/core/src/data-providers/json-entity-opfs-storage.ts
@@ -1,0 +1,24 @@
+import type { JsonEntityStorage } from './json-data-provider.js'
+
+export class JsonEntityOpfsStorage implements JsonEntityStorage {
+  //@internal
+  opfsRoot?: FileSystemDirectoryHandle
+  async getItem(entityDbName: string) {
+    if (!this.opfsRoot) {
+      this.opfsRoot = await navigator.storage.getDirectory()
+    }
+    const opfsFile = await this.opfsRoot.getFileHandle(entityDbName + '.json', {
+      create: true,
+    })
+    const readable = await opfsFile.getFile()
+    return await readable.text()
+  }
+  async setItem(entityDbName: string, json: string) {
+    const opfsFile = await this.opfsRoot.getFileHandle(entityDbName + '.json', {
+      create: true,
+    })
+    const writable = await opfsFile.createWritable()
+    await writable.write(json)
+    await writable.close()
+  }
+}