+++ projects/tests/tests/basicRowFunctionality.spec.ts
@@ -2137,7 +2137,7 @@ export class entityWithValidationsOnEntityEvent extends EntityBase {
   name: string
 }
 @Entity<EntityWithLateBoundDbName>('stam', {
-  sqlExpression: async (t) => '(select ' + t.fields.id.options.dbName + ')',
+  sqlExpression: async (t) => '(select ' + t.fields.id.dbName + ')',
 })
 export class EntityWithLateBoundDbName extends EntityBase {
   @Fields.integer({ dbName: 'CategoryID' })